+++ src/main/java/christmas/domain/orders/Orders.java
@@ -1,6 +1,5 @@
 package christmas.domain.orders;
 
-import christmas.domain.menu.MenuCategory;
 import christmas.validationUtils.OrdersValidationUtils;
 
 import java.util.List;
