+++ src/test/java/christmas/domain/event/EventDetailTest.java
@@ -1,6 +1,6 @@
 package christmas.domain.event;
 
-import christmas.domain.menu.MenuItem;
+import christmas.domain.orders.MenuItem;
 import christmas.domain.orders.OrderItem;
 import christmas.domain.orders.Orders;
 import christmas.domain.visitingDate.VisitingDate;
